Move vimwikigraph code outside of main.
